# openapi.yaml
openapi: 3.0.3
info:
  title: $7 Bootstrap Protocol — Agentic Commerce with x402 (Mock Payment)
  version: 0.1.0
  description: |
    Minimal local-only pay-to-deliver demo.

    Flow:
    Request → Invoice/Challenge → (Mock) Payment Verified → Deliver
servers:
  - url: http://localhost:3000
paths:
  /request:
    post:
      summary: Create an order (request a deliverable)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                  example: quant_snapshot_5m
                input:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Order created
  /invoice/{orderId}:
    get:
      summary: Get an invoice and x402-style challenge
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice/challenge
        '404':
          description: Order not found
  /verify/{orderId}:
    post:
      summary: Verify payment (mock)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                proof:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Payment verified
        '400':
          description: Must call invoice first
        '404':
          description: Order not found
  /deliver/{orderId}:
    post:
      summary: Deliver the requested artifact (requires paid)
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delivered
        '402':
          description: Payment required
        '404':
          description: Order not found